// src/types/cloudFormation.d.ts
type WafActionKeys = 'Allow' | 'Block';
type WafAction = WafActionKeys | CfnWafAction;
// ...
export type IntrinsictFunction = FnGetAtt | FnJoin | FnRef | FnSub;
// ...
type CfnWafAction =
  | { Allow: Record<string, never> }
  | { Deny: Record<string, never> };

type CfnWafRule = {
  Action?: CfnWafAction;
  Name: string;
  OverrideAction?: WafAction;
  Priority?: number;
  Statement: CfnWafRuleStatement;
  VisibilityConfig: unknown;
};

type CfnWafRuleStatement = {
  RateBasedStatement?: CfnWafRuleRateBasedStatement;
  ByteMatchStatement?: CfnWafRuleRateByteMatchStatement;
  AndStatement?: {
    Statements: CfnWafRuleStatement[];
  };
};

type CfnWafRuleRateBasedStatement = {
  AggregateKeyType: string;
  ForwardedIPConfig?: {
    FallbackBehavior: string;
    HeaderName: string;
  };
  Limit: number;
  ScopeDownStatement?: CfnWafRuleStatement;
};

type CfnWafRuleRateByteMatchStatement = {
  FieldToMatch: unknown;
  PositionalConstraint: string;
  SearchString?: string | IntrinsictFunction;
  SearchStringBase64?: string;
  TextTransformations: {
    Priority: number;
    Type: string;
  }[];
};